frontend/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
aif360==0.6.1
httpx==0.28.1
google-generativeai==0.7.0
python-multipart==0.0.9D